Complete Windmill construct review and reject malformed vault keys
Restore the feature-specific Velvet helper for optional Miru compilation. Preserve item-level API style justifications and existing task serialization.

Related: https://github.com/sequentech/meta/issues/10548

+++ packages/windmill/src/services/join.rs
@@ -448,7 +448,7 @@ mod tests {
     #[test]
     fn test_large_scale_auditable_count() -> Result<()> {
         const TOTAL_ENTRIES: u64 = 500;
-        const EXPECTED_AUDITABLE_COUNT: u64 = (TOTAL_ENTRIES / 2) as u64; // We will add only even users, so odds are auditable.
+        const EXPECTED_AUDITABLE_COUNT: u64 = TOTAL_ENTRIES / 2; // We will add only even users, so odds are auditable.
 
         let mut ballots_csv = String::new();
         let mut users_csv = String::new();
